@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Linq;
 using System.Reflection;
 using CQRSlite.Caching;
@@ -10,11 +11,11 @@ using CQRSlite.Routing;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using NSwag.AspNetCore;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 using RTTicTacToe.CQRS.Database;
 using RTTicTacToe.CQRS.Hubs;
 using RTTicTacToe.CQRS.ReadModel.Infrastructure;
@@ -51,6 +52,8 @@ namespace RTTicTacToe.WebApi
         /// <param name="services">Services.</param>
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddHttpContextAccessor();
+
             services.AddMemoryCache();
 
             //Add Cqrs services
@@ -71,7 +74,8 @@ namespace RTTicTacToe.WebApi
             //Scan for commandhandlers and eventhandlers
             services.Scan(scan => scan
                 .FromAssemblies(typeof(GameCommandHandlers).GetTypeInfo().Assembly)
-                .AddClasses(classes => classes.Where(x => {
+                .AddClasses(classes => classes.Where(x =>
+                {
                     var allInterfaces = x.GetInterfaces();
                     return
                         allInterfaces.Any(y => y.GetTypeInfo().IsGenericType && y.GetTypeInfo().GetGenericTypeDefinition() == typeof(IHandler<>)) ||
@@ -81,21 +85,36 @@ namespace RTTicTacToe.WebApi
                 .WithTransientLifetime()
             );
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-
-            services.AddSwaggerDocument(settings =>
+            services.AddSwaggerGen(c =>
             {
-                settings.Title = "RTTicTacToe - API";
+                c.SwaggerDoc("v1", new OpenApiInfo
+                {
+                    Version = "v1",
+                    Title = "RTTicTacToe - API",
+                    Description = "ASP.NET Core Web API with the implementation of the RTTicTacToe game backend",
+                    Contact = new OpenApiContact
+                    {
+                        Name = "José Pereira",
+                        Email = "leaopereira@gmail.com",
+                        Url = new Uri("https://twitter.com/z_leao")
+                    }
+                });
+
+                // Set the comments path for the Swagger JSON and UI.
+                var xmlPath = Path.Combine(AppContext.BaseDirectory, $"{Assembly.GetExecutingAssembly().GetName().Name}.xml");
+                c.IncludeXmlComments(xmlPath);
             });
 
             //Register routes
-            services.AddHttpContextAccessor(); // No longer registered by default in ASP.NET Core 2.1
             var serviceProvider = services.BuildServiceProvider();
             var registrar = new RouteRegistrar(new Provider(serviceProvider));
             registrar.RegisterInAssemblyOf(typeof(GameCommandHandlers));
 
             //Add support for SignalR
             services.AddSignalR();
+
+            // Configure controllers behaviour
+            services.AddControllers();
         }
 
         /// <summary>
@@ -104,30 +123,48 @@ namespace RTTicTacToe.WebApi
         /// </summary>
         /// <param name="app">App.</param>
         /// <param name="env">Env.</param>
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
-            
-            app.UseStaticFiles();
 
-            //Register SignalR endpoint
-            app.UseSignalR(routes =>
+            // Enable middleware to serve generated Swagger as a JSON endpoint.
+            app.UseSwagger();
+
+            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.),
+            // specifying the Swagger JSON endpoint.
+            app.UseSwaggerUI(c =>
             {
+                c.SwaggerEndpoint("/swagger/v1/swagger.json", "RTTicTacToe V1");
+            });
+
+            //app.UseStaticFiles();
+
+            app.UseHttpsRedirection();
+
+            app.UseRouting();
+
+            app.UseAuthorization();
+
+            app.UseEndpoints(routes =>
+            {
+                routes.MapControllers();
+
+                //Register SignalR endpoint
                 routes.MapHub<GameHub>("/api/gamehub");
             });
 
-            // Register the Swagger generator and the Swagger UI middlewares
-            app.UseSwagger(); 
-            app.UseSwaggerUi3();
-
-            app.UseMvc();
+            //Ensure the database is created, using migrations
+            using (var serviceScope = app.ApplicationServices.GetService<IServiceScopeFactory>().CreateScope())
+            {
+                var context = serviceScope.ServiceProvider.GetRequiredService<DatabaseContext>();
+                context.Database.Migrate();
+            }
         }
     }
 
-    //
     /// <summary>
     /// Provider that makes scoped services work inside router.
     /// </summary>
