@@ -8,7 +8,7 @@ namespace RTTicTacToe.CQRS.ReadModel.Dtos
         public string Name { get; private set; }
         public PlayerDto Player1 { get; set; }
         public PlayerDto Player2 { get; set; }
-        public int[,] Board { get; set; }
+        public int[][] Board { get; set; }
         public PlayerDto Winner { get; set; }
         public DateTime CreationDate { get; set; }
         public DateTime LastChangeDate { get; set; }
@@ -19,7 +19,7 @@ namespace RTTicTacToe.CQRS.ReadModel.Dtos
             Id = id;
             Name = name;
             Version = version;
-            Board = new int[3,3];
+            Board = new int[][] { new int[3], new int[3] };
         }
     }
 }
