@@ -1,11 +1,7 @@
-﻿using System;
-using System.IO;
-using Microsoft.AspNetCore;
+﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using RTTicTacToe.CQRS.Database;
+using Microsoft.Extensions.Hosting;
+using System.IO;
 
 namespace RTTicTacToe.WebApi
 {
@@ -20,25 +16,7 @@ namespace RTTicTacToe.WebApi
         /// <param name="args">The command-line arguments.</param>
         public static void Main(string[] args)
         {
-            //CreateWebHostBuilder(args).Build().Run();
-            var host = CreateWebHostBuilder(args).Build();
-
-            using (var scope = host.Services.CreateScope())
-            {
-                var services = scope.ServiceProvider;
-
-                try
-                {
-                    var context = services.GetRequiredService<DatabaseContext>();
-                    context.Database.Migrate();
-                }
-                catch (Exception ex)
-                {
-                    var logger = services.GetRequiredService<ILogger<Program>>();
-                    logger.LogError(ex, "An error occurred migrating the DB.");
-                }
-            }
-            host.Run();
+            CreateWebHostBuilder(args).Build().Run();
         }
 
         /// <summary>
@@ -49,7 +27,6 @@ namespace RTTicTacToe.WebApi
         public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
                 .UseStartup<Startup>()
-                .UseKestrel()
                 .UseContentRoot(Directory.GetCurrentDirectory());
     }
 }
